change(i3run): move app to at least workspace 5+ when sending it "away"
I like having workspaces 1-4 available for applications I don't have
bound to a dedicated key, so it's annoying when say vscode or my
main terminal with tmux panes get sent to workspaces 1-4.
This change the behavior of hidewindow.sh so every time a
workspace is sent "away" it will be sent to the next free workspace
UNLESS the next free workspace is < 5.